refactor: set up :common module and lift Nicephore constants/logger into it
- :common now compiles as a plain Kotlin library (no fabric-loom) with
  com.mojang:logging:1.6.11 compileOnly from libraries.minecraft.net.
  fabric-loom was removed because Mojang does not publish ProGuard mappings
  for MC 26.1, making officialMojangMappings() unavailable; the common module
  only needs LogUtils, not remapped Minecraft gameplay classes.
- Common Nicephore object holds MODID, MOD_NAME, LOGGER, and backgroundScope
  (TrashManager cleanup wired in a later task when the class moves to :common).
- :neoforge entry point renamed to NicephoreNeoForge.kt and simplified to
  delegate to Nicephore.startBackgroundTasks().

No behavioural change.
