+++ common/build.gradle
@@ -1,4 +1,24 @@
-// :common will be fleshed out in Task 4+.
-// For now, it is an empty Kotlin library so the multi-module tree resolves.
+plugins {
+    id 'net.neoforged.moddev' version '2.0.141'
+    id 'org.jetbrains.kotlin.plugin.serialization' version '2.3.20'
+}
+
+base {
+    archivesName = "${mod_id}-common"
+}
+
+// Apply NeoForge moddev in "library mode" — provides deobfuscated Minecraft classes
+// on the compile classpath without producing a mod jar or configuring NeoForge runs.
+// fabric-loom + officialMojangMappings() does not work for MC 26.1 (no ProGuard mappings published).
+neoForge {
+    version = project.neo_version
+}
+
 dependencies {
+    // Kotlin stdlib + kotlinx libs are provided by the loader at runtime (fabric-language-kotlin bundles,
+    // or NeoForge jarJars them). Here they are compileOnly so :common can reference them.
+    compileOnly 'org.jetbrains.kotlin:kotlin-stdlib:2.3.20'
+    compileOnly 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.10.2'
+    compileOnly 'org.jetbrains.kotlinx:kotlinx-datetime:0.7.1'
+    compileOnly 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.7.3'
 }
