+++ neoforge/build.gradle
@@ -48,6 +48,9 @@ configurations {
 }
 
 dependencies {
+    // Shared multi-loader module
+    implementation project(':common')
+
     implementation 'com.profesorfalken:jPowerShell:3.1.1'
     jarJar('com.profesorfalken:jPowerShell:[3.1.1,4.0.0)') {
         version { prefer '3.1.1' }
